backend/requirements.txt
fastapi
yfinance
requests
beautifulsoup4
python-dotenv
uvicorn
groq
autogen
plotly
lxml
pandas
numpy